functional oauth using github, parsing through user data and storing it in an object, oauth code is in seperate file from server
